[warnings]: Fixed some warnings

+++ source/libs/src_test/abc_backlight_controller/fake_abc_io_service.c
@@ -58,7 +58,7 @@ abc_ioService_read(int *const restrict pValue,
         return false;
     }
 
-    static const int s_MAX_FILENAME_LEN = 32;
+    static const size_t s_MAX_FILENAME_LEN = 32;
 
     assert(pValue &&
            pFileName &&
